Update commands.js
I fixed the problems you mentioned, and added the line you mentioned after it selects the final choice. I took this.say(retort) out of every if, gave it the right amount of arguments, and moved it to the final conditional instead. I also made snen spam the last thing it checks for since it's less likely to happen than caps or stretching.