Added .modchat for drivers.
I implemented a command strictly for a moderator ranked bot. The commands newly added are .modchatac, .modchatv and .modchatoff. These 3 commands allow drivers to turn on modchat if a Room Owner or moderator is not active. These commands I think will help room auth to allow themselves access and not contacting a global mod to just turn on mod chat.