Source Command
Hi, i want to do a pull request for this command, is just a source command i mean if someone want to see the bot source just do .source on pm and bot give the link of github in that case will be https://github.com/TalkTakesTime/Pokemon-Showdown-Bot i hope you like the idea! 

Thanks.